Remove StatusListTokenValidator
